Create README.md
RSSI-based indoor localization system developed in this project utilizes wireless sensor nodes such as access points (APs) or any other sensor modalities that are capable of generating a Receive Signal Strength Indication (RSSI), which can be used to determine the location of a target object in an indoor environment. A device with an RFID tag or a smartphone with a wireless antenna scans the RSSIs from surrounding active APs, and sends the recorded values to the server using HTTP communication protocol. The ‘Localization
service’ deployed on the server receives the RSSIs, extracts location related information and delivers to the component that has similarity matching algorithmic implementation. A room or an area with the most similar RSSIs from the database that has pre-gathered RSSI samples stored in is returned to the client/visitor as a current estimated location.